fix clients
